+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Removed `isRequired` for `acceptAction` and added custom validation for it.
 
 3.21.0 - (December 10, 2019)
 ------------------